refactor: update the games page
