--- src/components/HomeComponent.jsx
@@ -1,9 +1,0 @@
-import React from react;
-
-function HomeComponent() {
-  return (
-    <div>Hello</div>
-  );
-}
-
-export default HomeComponent;
